Put Animated list widget in separate file
