+++ app/lib/features/chat/widgets/chats_list.dart
@@ -1,4 +1,3 @@
-import 'package:acter/common/extensions/options.dart';
 import 'package:acter/common/providers/chat_providers.dart';
 import 'package:acter/common/providers/common_providers.dart';
 import 'package:acter/common/toolkit/buttons/primary_action_button.dart';
@@ -6,8 +5,7 @@ import 'package:acter/common/utils/routes.dart';
 import 'package:acter/common/widgets/empty_state_widget.dart';
 import 'package:acter/features/chat/providers/chat_providers.dart';
 import 'package:acter/features/chat/providers/room_list_filter_provider.dart';
-import 'package:acter/features/chat_ui_showcase/widgets/chat_item_widget.dart';
-import 'package:diffutil_dart/diffutil.dart';
+import 'package:acter/features/chat/widgets/animated_chats_list_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:acter/l10n/generated/l10n.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -90,142 +88,6 @@ class ChatsList extends ConsumerWidget {
   }
 
   Widget _renderList(BuildContext context, List<String> chats) {
-    return AnimatedChatsList(entries: chats, onSelected: onSelected);
-  }
-}
-
-class AnimatedChatsList extends StatefulWidget {
-  final Function(String)? onSelected;
-  final List<String> entries;
-
-  const AnimatedChatsList({super.key, required this.entries, this.onSelected});
-
-  @override
-  AnimatedChatsListState createState() => AnimatedChatsListState();
-}
-
-class AnimatedChatsListState extends State<AnimatedChatsList> {
-  late GlobalKey<AnimatedListState> _listKey;
-  late List<String> _currentList;
-
-  @override
-  void initState() {
-    super.initState();
-    _reset();
-  }
-
-  void _reset() {
-    _listKey = GlobalKey<AnimatedListState>(debugLabel: 'chat rooms list');
-    _currentList = List.of(widget.entries);
-  }
-
-  @override
-  void didUpdateWidget(AnimatedChatsList oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (_listKey.currentState == null) {
-      _log.fine('no state, hard reset');
-      // we can ignore the diffing as we aren't live, just reset
-      setState(_reset);
-      return;
-    } else {
-      refreshList();
-    }
-  }
-
-  void refreshList() {
-    _log.fine('refreshing');
-    WidgetsBinding.instance.addPostFrameCallback((Duration duration) {
-      _log.fine('diffing $_currentList, ${widget.entries}');
-      final diffResult = calculateListDiff<String>(
-        _currentList,
-        widget.entries,
-        detectMoves: false,
-      );
-      for (final update in diffResult.getUpdatesWithData()) {
-        update.when(
-          insert: _insert,
-          remove: _remove,
-          change: (pos, oldData, newData) {
-            _remove(pos, oldData);
-            _insert(pos, newData);
-          },
-          move: (from, to, item) {
-            _remove(from, item);
-            _insert(to, item);
-          },
-        );
-      }
-      _log.fine('done diffing');
-    });
-  }
-
-  void _insert(int pos, String data) {
-    _log.fine('insert $pos: $data');
-    _currentList.insert(pos, data);
-    _listKey.currentState.map(
-      (curState) =>
-          curState.insertItem(pos, duration: const Duration(milliseconds: 300)),
-      orElse: () => _log.fine('we are not'),
-    );
-  }
-
-  void _remove(int pos, String data) {
-    _currentList.removeAt(pos);
-    _listKey.currentState.map(
-      (curState) => curState.removeItem(
-        pos,
-        (context, animation) => _removedItemBuilder(data, context, animation),
-      ),
-      orElse: () => _log.fine('we are not'),
-    );
-  }
-
-  Widget _removedItemBuilder(
-    String roomId,
-    BuildContext context,
-    Animation<double> animation,
-  ) {
-    return ChatItemWidget(
-      animation: animation,
-      key: Key('convo-card-$roomId-removed'),
-      roomId: roomId,
-      onTap: () {
-        final onSelected = widget.onSelected;
-        if (onSelected != null) onSelected(roomId);
-      },
-    );
-  }
-
-  Widget buildItem(
-    BuildContext context,
-    int index,
-    Animation<double> animation,
-  ) {
-    if (index >= _currentList.length) {
-      // due to this updating with animations, we sometimes run out of the index
-      // while still manipulating it. Ignore that case with just some empty boxes.
-      return const SizedBox.shrink();
-    }
-    final roomId = _currentList[index];
-    _log.fine('render $roomId');
-    return ChatItemWidget(
-      animation: animation,
-      key: Key('convo-card-$roomId'),
-      roomId: roomId,
-      onTap: () {
-        final onSelected = widget.onSelected;
-        if (onSelected != null) onSelected(roomId);
-      },
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    _log.fine('render list $_currentList');
-    return AnimatedList(
-      key: _listKey,
-      initialItemCount: _currentList.length,
-      itemBuilder: buildItem,
-    );
+    return AnimatedChatsListWidget(entries: chats, onSelected: onSelected);
   }
 }
