+++ app/lib/features/chat_ui_showcase/pages/chat_list_showcase_page.dart
@@ -1,5 +1,5 @@
 import 'package:acter/common/widgets/plus_icon_widget.dart';
-import 'package:acter/features/chat/widgets/chats_list.dart';
+import 'package:acter/features/chat/widgets/animated_chats_list_widget.dart';
 import 'package:acter/features/chat_ui_showcase/models/convo_showcase_list.dart';
 import 'package:acter/l10n/generated/l10n.dart';
 import 'package:flutter/material.dart';
@@ -15,7 +15,7 @@ class ChatListShowcasePage extends StatelessWidget {
         title: Text(lang.chat),
         actions: [PlusIconWidget(onPressed: () {})],
       ),
-      body: AnimatedChatsList(
+      body: AnimatedChatsListWidget(
         entries: mockChatList.map((e) => e.roomId).toList(),
       ),
     );
